Introduce JSON service
