+++ src/App.svelte
@@ -5,6 +5,7 @@
 	import { CloningService } from "./services/cloningService";
 	import { CommandService } from "./services/commandService";
 	import { GridService } from "./services/gridService";
+	import { JSONService } from "./services/jsonService";
 	import { LocalStorageService } from "./services/localStorageService";
 	import { PaneService } from "./services/paneService";
 	import { TabService } from "./services/tabService";
@@ -16,7 +17,8 @@
 	const tabService = new TabService();
 	const commandService = new CommandService(paneService);
 	const uiService = new UIService(gridService, commandService);
-	const localStorageService = new LocalStorageService(paneService);
+	const jsonService = new JSONService(paneService);
+	const localStorageService = new LocalStorageService(jsonService);
 
 	let layouts = localStorageService.retrieveLayouts();
